+++ src/test/kotlin/io/kstuff/log/test/LoggerUtilTest.kt
@@ -2,7 +2,7 @@
  * @(#) LoggerUtilTest.kt
  *
  * log-front-kotlin  Logging interface in Kotlin
- * Copyright (c) 2020, 2022, 2024 Peter Wall
+ * Copyright (c) 2020, 2022, 2024, 2925 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,7 +23,7 @@
  * SOFTWARE.
  */
 
-package net.pwall.log.test
+package io.kstuff.log.test
 
 import kotlin.test.Test
 
@@ -37,49 +37,70 @@ import io.kstuff.test.shouldBeSameInstance
 import io.kstuff.test.shouldContain
 import io.kstuff.test.shouldThrow
 
-import net.pwall.log.DynamicLoggerFactory
-import net.pwall.log.Level
-import net.pwall.log.Log
-import net.pwall.log.LogItem
-import net.pwall.log.LogList
-import net.pwall.log.Logger
-import net.pwall.log.assertHasDebug
-import net.pwall.log.assertHasError
-import net.pwall.log.assertHasInfo
-import net.pwall.log.assertHasTrace
-import net.pwall.log.assertHasWarning
-import net.pwall.log.getLogger
-import net.pwall.log.isDebug
-import net.pwall.log.isError
-import net.pwall.log.isInfo
-import net.pwall.log.isTrace
-import net.pwall.log.isWarning
-import net.pwall.log.subList
+import io.jstuff.log.DynamicLoggerFactory
+import io.jstuff.log.Level
+import io.jstuff.log.Log
+import io.jstuff.log.LogItem
+import io.jstuff.log.LogList
+import io.jstuff.log.Logger
+
+import io.kstuff.log.shouldHaveDebug
+import io.kstuff.log.shouldHaveDebugMatching
+import io.kstuff.log.shouldHaveError
+import io.kstuff.log.shouldHaveErrorMatching
+import io.kstuff.log.shouldHaveInfo
+import io.kstuff.log.shouldHaveInfoMatching
+import io.kstuff.log.shouldHaveTrace
+import io.kstuff.log.shouldHaveTraceMatching
+import io.kstuff.log.shouldHaveWarning
+import io.kstuff.log.shouldHaveWarningMatching
+import io.kstuff.log.getLogger
+import io.kstuff.log.isDebug
+import io.kstuff.log.isDebugMatching
+import io.kstuff.log.isDebugContaining
+import io.kstuff.log.isError
+import io.kstuff.log.isErrorMatching
+import io.kstuff.log.isErrorContaining
+import io.kstuff.log.isInfo
+import io.kstuff.log.isInfoMatching
+import io.kstuff.log.isInfoContaining
+import io.kstuff.log.isTrace
+import io.kstuff.log.isTraceMatching
+import io.kstuff.log.isTraceContaining
+import io.kstuff.log.isWarning
+import io.kstuff.log.isWarningMatching
+import io.kstuff.log.isWarningContaining
+import io.kstuff.log.shouldHaveDebugContaining
+import io.kstuff.log.shouldHaveErrorContaining
+import io.kstuff.log.shouldHaveInfoContaining
+import io.kstuff.log.shouldHaveTraceContaining
+import io.kstuff.log.shouldHaveWarningContaining
+import io.kstuff.log.subList
 
 class LoggerUtilTest {
 
     @Test fun `should get logger for current class`() {
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger()) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Log.getDefaultLoggerFactory().defaultLevel
             clock shouldBe Log.getDefaultLoggerFactory().defaultClock
         }
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(Level.DEBUG)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Level.DEBUG
             clock shouldBe Log.getDefaultLoggerFactory().defaultClock
         }
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(fixedClock)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Log.getDefaultLoggerFactory().defaultLevel
             clock shouldBeSameInstance fixedClock
         }
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(Level.DEBUG, fixedClock)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Level.DEBUG
             clock shouldBeSameInstance fixedClock
         }
@@ -115,25 +136,25 @@ class LoggerUtilTest {
     @Test fun `should get logger for Java Class`() {
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(LoggerUtilTest::class.java)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Log.getDefaultLoggerFactory().defaultLevel
             clock shouldBe Log.getDefaultLoggerFactory().defaultClock
         }
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(LoggerUtilTest::class.java, Level.DEBUG)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Level.DEBUG
             clock shouldBe Log.getDefaultLoggerFactory().defaultClock
         }
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(LoggerUtilTest::class.java, fixedClock)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Log.getDefaultLoggerFactory().defaultLevel
             clock shouldBeSameInstance fixedClock
         }
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(LoggerUtilTest::class.java, Level.DEBUG, fixedClock)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Level.DEBUG
             clock shouldBeSameInstance fixedClock
         }
@@ -142,25 +163,25 @@ class LoggerUtilTest {
     @Test fun `should get logger for KClass`() {
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(LoggerUtilTest::class)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Log.getDefaultLoggerFactory().defaultLevel
             clock shouldBe Log.getDefaultLoggerFactory().defaultClock
         }
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(LoggerUtilTest::class, Level.DEBUG)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Level.DEBUG
             clock shouldBe Log.getDefaultLoggerFactory().defaultClock
         }
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(LoggerUtilTest::class, fixedClock)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Log.getDefaultLoggerFactory().defaultLevel
             clock shouldBeSameInstance fixedClock
         }
         Log.setDefaultLoggerFactory(DynamicLoggerFactory())
         with(getLogger(LoggerUtilTest::class, Level.DEBUG, fixedClock)) {
-            name shouldBe "net.pwall.log.test.LoggerUtilTest"
+            name shouldBe "io.kstuff.log.test.LoggerUtilTest"
             level shouldBe Level.DEBUG
             clock shouldBeSameInstance fixedClock
         }
@@ -169,67 +190,86 @@ class LoggerUtilTest {
     @Test fun `should get Logger from LoggerFactory by KClass`() {
         val factory = Log.getDefaultLoggerFactory()
         val logger = factory.getLogger(LoggerUtilTest::class)
-        logger.name shouldBe "net.pwall.log.test.LoggerUtilTest"
+        logger.name shouldBe "io.kstuff.log.test.LoggerUtilTest"
     }
     @Test fun `should get logger in companion object`() {
-        companionObjectLogger1.name shouldBe "net.pwall.log.test.LoggerUtilTest"
-        companionObjectLogger2.name shouldBe "net.pwall.log.test.LoggerUtilTest"
+        companionObjectLogger1.name shouldBe "io.kstuff.log.test.LoggerUtilTest"
+        companionObjectLogger2.name shouldBe "io.kstuff.log.test.LoggerUtilTest"
     }
 
     @Test fun `should check whether LogItem is trace`() {
         val logItem = LogItem(System.currentTimeMillis(), "xyz", Level.TRACE, "Hello", null)
         logItem isTrace "Hello" shouldBe true
-        logItem isTrace Regex("^H") shouldBe true
+        logItem isTraceContaining  "Hell" shouldBe true
+        logItem isTraceMatching Regex("^H") shouldBe true
         logItem isDebug "Hello" shouldBe false
         logItem isTrace "Goodbye" shouldBe false
-        logItem isTrace Regex("^Hello world") shouldBe false
+        logItem isTraceMatching Regex("^Hello world") shouldBe false
     }
 
     @Test fun `should check whether LogItem is debug`() {
         val logItem = LogItem(System.currentTimeMillis(), "xyz", Level.DEBUG, "Hello", null)
         logItem isDebug "Hello" shouldBe true
-        logItem isDebug Regex("^H") shouldBe true
+        logItem isDebugContaining  "ell" shouldBe true
+        logItem isDebugMatching Regex("^H") shouldBe true
         logItem isInfo "Hello" shouldBe false
         logItem isDebug "Goodbye" shouldBe false
-        logItem isDebug Regex("^Hello world") shouldBe false
+        logItem isDebugMatching Regex("^Hello world") shouldBe false
     }
 
     @Test fun `should check whether LogItem is info`() {
         val logItem = LogItem(System.currentTimeMillis(), "xyz", Level.INFO, "Hello", null)
         logItem isInfo "Hello" shouldBe true
-        logItem isInfo Regex("^H") shouldBe true
+        logItem isInfoContaining  "He" shouldBe true
+        logItem isInfoMatching Regex("^H") shouldBe true
         logItem isWarning "Hello" shouldBe false
         logItem isInfo "Goodbye" shouldBe false
-        logItem isInfo Regex("^Hello world") shouldBe false
+        logItem isInfoMatching Regex("^Hello world") shouldBe false
     }
 
     @Test fun `should check whether LogItem is warning`() {
         val logItem = LogItem(System.currentTimeMillis(), "xyz", Level.WARN, "Hello", null)
         logItem isWarning "Hello" shouldBe true
-        logItem isWarning Regex("^H") shouldBe true
+        logItem isWarningContaining  "lo" shouldBe true
+        logItem isWarningMatching Regex("^H") shouldBe true
         logItem isError "Hello" shouldBe false
         logItem isWarning "Goodbye" shouldBe false
-        logItem isWarning Regex("^Hello world") shouldBe false
+        logItem isWarningMatching Regex("^Hello world") shouldBe false
     }
 
     @Test fun `should check whether LogItem is error`() {
         val logItem = LogItem(System.currentTimeMillis(), "xyz", Level.ERROR, "Hello", null)
         logItem isError "Hello" shouldBe true
-        logItem isError Regex("^H") shouldBe true
+        logItem isErrorContaining  "el" shouldBe true
+        logItem isErrorMatching Regex("^H") shouldBe true
         logItem isWarning  "Hello" shouldBe false
         logItem isError "Goodbye" shouldBe false
-        logItem isError Regex("^Hello world") shouldBe false
+        logItem isErrorMatching Regex("^Hello world") shouldBe false
     }
 
     @Test fun `should check whether LogList contains TRACE item`() {
         LogList().use {  list ->
             val logger = getLogger("LogTrace", Level.TRACE)
             logger.trace { "Trace message" }
-            list.assertHasTrace("Trace message")
-            shouldThrow<AssertionError> { list.assertHasTrace("Wrong") }.let {
+            list.shouldHaveTrace("Trace message")
+            shouldThrow<AssertionError> { list.shouldHaveTrace("Wrong") }.let {
                 it.message.let {  message ->
                     message.shouldBeNonNull()
                     message shouldContain "LogList does not contain TRACE Wrong"
+                }
+            }
+        }
+    }
+
+    @Test fun `should check whether LogList contains TRACE item containing string`() {
+        LogList().use {  list ->
+            val logger = getLogger("LogTrace", Level.TRACE)
+            logger.trace { "Trace message" }
+            list.shouldHaveTraceContaining("mess")
+            shouldThrow<AssertionError> { list.shouldHaveTraceContaining("Wrong") }.let {
+                it.message.let {  message ->
+                    message.shouldBeNonNull()
+                    message shouldContain "LogList does not contain TRACE containing Wrong"
                 }
             }
         }
@@ -239,8 +279,8 @@ class LoggerUtilTest {
         LogList().use {  list ->
             val logger = getLogger("LogTrace", Level.TRACE)
             logger.trace { "Trace message" }
-            list.assertHasTrace(Regex("Trace"))
-            shouldThrow<AssertionError> { list.assertHasTrace(Regex("Wrong")) }.let {
+            list.shouldHaveTraceMatching(Regex("Trace"))
+            shouldThrow<AssertionError> { list.shouldHaveTraceMatching(Regex("Wrong")) }.let {
                 it.message.let {  message ->
                     message.shouldBeNonNull()
                     message shouldContain "LogList does not contain TRACE Wrong"
@@ -253,11 +293,25 @@ class LoggerUtilTest {
         LogList().use {  list ->
             val logger = getLogger("LogDebug", Level.DEBUG)
             logger.debug { "Debug message" }
-            list.assertHasDebug("Debug message")
-            shouldThrow<AssertionError> { list.assertHasDebug("Wrong") }.let {
+            list.shouldHaveDebug("Debug message")
+            shouldThrow<AssertionError> { list.shouldHaveDebug("Wrong") }.let {
                 it.message.let {  message ->
                     message.shouldBeNonNull()
                     message shouldContain "LogList does not contain DEBUG Wrong"
+                }
+            }
+        }
+    }
+
+    @Test fun `should check whether LogList contains DEBUG item containing string`() {
+        LogList().use {  list ->
+            val logger = getLogger("LogDebug", Level.DEBUG)
+            logger.debug { "Debug message" }
+            list.shouldHaveDebugContaining("bug")
+            shouldThrow<AssertionError> { list.shouldHaveDebugContaining("Wrong") }.let {
+                it.message.let {  message ->
+                    message.shouldBeNonNull()
+                    message shouldContain "LogList does not contain DEBUG containing Wrong"
                 }
             }
         }
@@ -267,8 +321,8 @@ class LoggerUtilTest {
         LogList().use {  list ->
             val logger = getLogger("LogDebug", Level.DEBUG)
             logger.debug { "Debug message" }
-            list.assertHasDebug(Regex("Debug"))
-            shouldThrow<AssertionError> { list.assertHasDebug(Regex("Wrong")) }.let {
+            list.shouldHaveDebugMatching(Regex("Debug"))
+            shouldThrow<AssertionError> { list.shouldHaveDebugMatching(Regex("Wrong")) }.let {
                 it.message.let {  message ->
                     message.shouldBeNonNull()
                     message shouldContain "LogList does not contain DEBUG Wrong"
@@ -280,8 +334,8 @@ class LoggerUtilTest {
     @Test fun `should check whether LogList contains INFO item`() {
         LogList().use {  list ->
             companionObjectLogger1.info { "Message" }
-            list.assertHasInfo("Message")
-            shouldThrow<AssertionError> { list.assertHasInfo("Wrong") }.let {
+            list.shouldHaveInfo("Message")
+            shouldThrow<AssertionError> { list.shouldHaveInfo("Wrong") }.let {
                 it.message.let {  message ->
                     message.shouldBeNonNull()
                     message shouldContain "LogList does not contain INFO Wrong"
@@ -290,11 +344,24 @@ class LoggerUtilTest {
         }
     }
 
+    @Test fun `should check whether LogList contains INFO item containing string`() {
+        LogList().use {  list ->
+            companionObjectLogger1.info { "Message" }
+            list.shouldHaveInfoContaining("sage")
+            shouldThrow<AssertionError> { list.shouldHaveInfoContaining("Wrong") }.let {
+                it.message.let {  message ->
+                    message.shouldBeNonNull()
+                    message shouldContain "LogList does not contain INFO containing Wrong"
+                }
+            }
+        }
+    }
+
     @Test fun `should check whether LogList contains INFO item matching Regex`() {
         LogList().use {  list ->
             companionObjectLogger1.info { "Message" }
-            list.assertHasInfo(Regex("Mess"))
-            shouldThrow<AssertionError> { list.assertHasInfo(Regex("Wrong")) }.let {
+            list.shouldHaveInfoMatching(Regex("Mess"))
+            shouldThrow<AssertionError> { list.shouldHaveInfoMatching(Regex("Wrong")) }.let {
                 it.message.let {  message ->
                     message.shouldBeNonNull()
                     message shouldContain "LogList does not contain INFO Wrong"
@@ -306,8 +373,8 @@ class LoggerUtilTest {
     @Test fun `should check whether LogList contains WARN item`() {
         LogList().use {  list ->
             companionObjectLogger1.warn { "Warning message" }
-            list.assertHasWarning("Warning message")
-            shouldThrow<AssertionError> { list.assertHasWarning("Wrong") }.let {
+            list.shouldHaveWarning("Warning message")
+            shouldThrow<AssertionError> { list.shouldHaveWarning("Wrong") }.let {
                 it.message.let {  message ->
                     message.shouldBeNonNull()
                     message shouldContain "LogList does not contain WARN Wrong"
@@ -316,11 +383,24 @@ class LoggerUtilTest {
         }
     }
 
+    @Test fun `should check whether LogList contains WARN item containing string`() {
+        LogList().use {  list ->
+            companionObjectLogger1.warn { "Warning message" }
+            list.shouldHaveWarningContaining("age")
+            shouldThrow<AssertionError> { list.shouldHaveWarningContaining("Wrong") }.let {
+                it.message.let {  message ->
+                    message.shouldBeNonNull()
+                    message shouldContain "LogList does not contain WARN containing Wrong"
+                }
+            }
+        }
+    }
+
     @Test fun `should check whether LogList contains WARN item matching Regex`() {
         LogList().use {  list ->
             companionObjectLogger1.warn { "Warning message" }
-            list.assertHasWarning(Regex("Warning"))
-            shouldThrow<AssertionError> { list.assertHasWarning(Regex("Wrong")) }.let {
+            list.shouldHaveWarningMatching(Regex("Warning"))
+            shouldThrow<AssertionError> { list.shouldHaveWarningMatching(Regex("Wrong")) }.let {
                 it.message.let {  message ->
                     message.shouldBeNonNull()
                     message shouldContain "LogList does not contain WARN Wrong"
@@ -332,8 +412,8 @@ class LoggerUtilTest {
     @Test fun `should check whether LogList contains ERROR item`() {
         LogList().use {  list ->
             companionObjectLogger1.error { "Error message" }
-            list.assertHasError("Error message")
-            shouldThrow<AssertionError> { list.assertHasError("Wrong") }.let {
+            list.shouldHaveError("Error message")
+            shouldThrow<AssertionError> { list.shouldHaveError("Wrong") }.let {
                 it.message.let {  message ->
                     message.shouldBeNonNull()
                     message shouldContain "LogList does not contain ERROR Wrong"
@@ -342,11 +422,24 @@ class LoggerUtilTest {
         }
     }
 
+    @Test fun `should check whether LogList contains ERROR item containing string`() {
+        LogList().use {  list ->
+            companionObjectLogger1.error { "Error message" }
+            list.shouldHaveErrorContaining("Err")
+            shouldThrow<AssertionError> { list.shouldHaveErrorContaining("Wrong") }.let {
+                it.message.let {  message ->
+                    message.shouldBeNonNull()
+                    message shouldContain "LogList does not contain ERROR containing Wrong"
+                }
+            }
+        }
+    }
+
     @Test fun `should check whether LogList contains ERROR item matching Regex`() {
         LogList().use {  list ->
             companionObjectLogger1.error { "Error message" }
-            list.assertHasError(Regex("Error"))
-            shouldThrow<AssertionError> { list.assertHasError(Regex("Wrong")) }.let {
+            list.shouldHaveErrorMatching(Regex("Error"))
+            shouldThrow<AssertionError> { list.shouldHaveErrorMatching(Regex("Wrong")) }.let {
                 it.message.let {  message ->
                     message.shouldBeNonNull()
                     message shouldContain "LogList does not contain ERROR Wrong"
@@ -359,7 +452,7 @@ class LoggerUtilTest {
         LogList().use {  list ->
             val logger = getLogger("TestLog")
             logger.info { "Message" }
-            list.subList("TestLog").assertHasInfo("Message")
+            list.subList("TestLog").shouldHaveInfo("Message")
         }
     }